requirements.txt
python>=3.11.4
kivy==2.2.1
kivymd==1.1.1
opencv-python
pyzbar
gspread
datetime
yagmail
google-auth
google-auth-oauthlib
google-auth-httplib2
camera4kivy
regex
httplib2
google-api-python-client
oauth2client
Pillow
pywin32
PyAutoGUI
